backend/upload-file/requirements.txt
boto3==1.34.96
PyJWT>=2.6.0